feat: Create the initial HTML file with it's content
Create the initial HTML file from scratch with all the page content,
header, sections and footer and give them classes to be used in styling
with CSS and interactivity with JavaScript.

link the HTML file with CSS file and JavaScript file.

Add comments to each section to indicate what does it refer to.