feat: Implement dynamic navbar, highlight sections
Create dynamic nav bar that scales with the number of sections on the page.

Implement feature to change the section background color when scrolled to or navigated to.

Add an event listener for the nav bar buttons to navigate to the required section.

Add an event listener for the scroll to highlight the section that has been scrolled to.